Fully implemented Banker Class.
Some accessors may be unnecessary. Could reduce code in lines 91-100 and 103-112 by combining if statements into one check. Chose not to in order to make code more clear. Not sure if player name needs to be printed in the upgrade statements. If so, then upgrade methods should be moved into Player Class as Banker Class does not have access to playerName attribute.